Switch the order of checks in node_ptr
First check if this node is on the freelist. If it is then we don't
need to check the node id (nid). Some significant percentage of the
time the node will be on the freelist and if it is we can return None
without checking the nodeId (nid) which is a bit faster. Given that
node_ptr is called on almost every method its helpful to make it as
efficient as possible.